migrate passed in files (or discover files)
