+++ packages/@tailwindcss-upgrade/package.json
@@ -30,10 +30,12 @@
   },
   "dependencies": {
     "enhanced-resolve": "^5.17.1",
+    "fast-glob": "^3.3.2",
     "mri": "^1.2.0",
     "picocolors": "^1.0.1",
     "postcss": "^8.4.41",
-    "postcss-import": "^16.1.0"
+    "postcss-import": "^16.1.0",
+    "tailwindcss": "workspace:^"
   },
   "devDependencies": {
     "@types/node": "catalog:",
