+++ packages/@tailwindcss-upgrade/tsup.config.ts
@@ -1,4 +1,4 @@
-import {defineConfig} from 'tsup'
+import { defineConfig } from 'tsup'
 
 export default defineConfig({
   format: ['esm'],
